backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
fivetran-connector-sdk
requests
elasticsearch
python-dotenv
google-cloud-aiplatform
pydantic
langchain==0.3.7
langchain-core==0.3.15
langchain-elasticsearch==0.3.2  # Released Jan 20, 2025
langchain-google-vertexai==2.1.2  # Released Sep 16, 2025
langchain-community==0.3.5
